+++ cmd/athenz-sia/main.go
@@ -10,8 +10,8 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/AthenZ/k8s-athenz-sia/pkg/log"
-	"github.com/AthenZ/k8s-athenz-sia/pkg/util"
+	"github.com/AthenZ/k8s-athenz-sia/third_party/log"
+	"github.com/AthenZ/k8s-athenz-sia/third_party/util"
 	"github.com/pkg/errors"
 
 	"github.com/AthenZ/k8s-athenz-sia/pkg/identity"
