+++ pkg/identity/tokend.go
@@ -11,8 +11,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/AthenZ/k8s-athenz-sia/pkg/log"
-	"github.com/AthenZ/k8s-athenz-sia/pkg/util"
+	"github.com/AthenZ/k8s-athenz-sia/third_party/log"
+	"github.com/AthenZ/k8s-athenz-sia/third_party/util"
 	"github.com/cenkalti/backoff"
 	"github.com/pkg/errors"
 )
